sitemate_challenge_backend/requirements.txt
alembic==1.0.8
python-dotenv==0.18.0
SQLAlchemy==1.4.20
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytest==4.3.1
pytest-cov~=2.7.1
Flask==2.2.2
Werkzeug==2.3.7
Flask-Cors==3.0.10
